// typescript2/pkg-baml-tooling/src/generated/tooling.ts
export interface ProjectLayout {
  configPath: string;
  roots: string[];
  sourceFiles: string[];
  watchFiles: string[];
}

export interface Capabilities {
  protocol: string;
  features: string[];
  compilerVersion: string;
}

export interface CheckResult {
  revision: number;
  diagnostics: Diagnostic[];
}

export interface Diagnostic {
  code: string;
  message: string;
  severity: string;
  location: Location | undefined;
}

export interface VirtualModule {
  id: string;
  code: string;
  declaration: string;
  map: SegmentMap | undefined;
  watchFiles: string[];
  fingerprint: string;
  revision: number;
  stale: boolean;
  runtimeId: string;
}

export interface RuntimeModule {
  id: string;
  code: string;
  watchFiles: string[];
  fingerprint: string;
  revision: number;
}
// ...
export interface Location {
  path: string;
  startUtf8: number;
  lengthUtf8: number;
}

export interface Locations {
  locations: Location[];
}

export interface Hover {
  markdown: string;
  location: Location | undefined;
  symbolId: string;
}

export interface CompletionItem {
  label: string;
  kind: string;
  detail: string;
  documentation: string;
  symbolId: string;
}

export interface Completions {
  items: CompletionItem[];
}

export interface TextEdit {
  location: Location | undefined;
  newText: string;
}

export interface WorkspaceEdit {
  edits: TextEdit[];
}
// ...
export const Capabilities: MessageFns<Capabilities> = {
  encode(message: Capabilities, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.protocol !== "") {
      writer.uint32(10).string(message.protocol);
    }
    for (const v of message.features) {
      writer.uint32(18).string(v!);
    }
    if (message.compilerVersion !== "") {
      writer.uint32(26).string(message.compilerVersion);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Capabilities {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCapabilities();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.protocol = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.features.push(reader.string());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.compilerVersion = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Capabilities {
    return {
      protocol: isSet(object.protocol) ? globalThis.String(object.protocol) : "",
      features: globalThis.Array.isArray(object?.features) ? object.features.map((e: any) => globalThis.String(e)) : [],
      compilerVersion: isSet(object.compilerVersion)
        ? globalThis.String(object.compilerVersion)
        : isSet(object.compiler_version)
        ? globalThis.String(object.compiler_version)
        : "",
    };
  },

  toJSON(message: Capabilities): unknown {
    const obj: any = {};
    if (message.protocol !== "") {
      obj.protocol = message.protocol;
    }
    if (message.features?.length) {
      obj.features = message.features;
    }
    if (message.compilerVersion !== "") {
      obj.compilerVersion = message.compilerVersion;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Capabilities>, I>>(base?: I): Capabilities {
    return Capabilities.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Capabilities>, I>>(object: I): Capabilities {
    const message = createBaseCapabilities();
    message.protocol = object.protocol ?? "";
    message.features = object.features?.map((e) => e) || [];
    message.compilerVersion = object.compilerVersion ?? "";
    return message;
  },
};

function createBaseCheckResult(): CheckResult {
  return { revision: 0, diagnostics: [] };
}

export const CheckResult: MessageFns<CheckResult> = {
  encode(message: CheckResult, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.revision !== 0) {
      writer.uint32(8).uint64(message.revision);
    }
    for (const v of message.diagnostics) {
      Diagnostic.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CheckResult {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCheckResult();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.revision = longToNumber(reader.uint64());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.diagnostics.push(Diagnostic.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CheckResult {
    return {
      revision: isSet(object.revision) ? globalThis.Number(object.revision) : 0,
      diagnostics: globalThis.Array.isArray(object?.diagnostics)
        ? object.diagnostics.map((e: any) => Diagnostic.fromJSON(e))
        : [],
    };
  },

  toJSON(message: CheckResult): unknown {
    const obj: any = {};
    if (message.revision !== 0) {
      obj.revision = Math.round(message.revision);
    }
    if (message.diagnostics?.length) {
      obj.diagnostics = message.diagnostics.map((e) => Diagnostic.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CheckResult>, I>>(base?: I): CheckResult {
    return CheckResult.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CheckResult>, I>>(object: I): CheckResult {
    const message = createBaseCheckResult();
    message.revision = object.revision ?? 0;
    message.diagnostics = object.diagnostics?.map((e) => Diagnostic.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const Hover: MessageFns<Hover> = {
  encode(message: Hover, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.markdown !== "") {
      writer.uint32(10).string(message.markdown);
    }
    if (message.location !== undefined) {
      Location.encode(message.location, writer.uint32(18).fork()).join();
    }
    if (message.symbolId !== "") {
      writer.uint32(26).string(message.symbolId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Hover {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseHover();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.markdown = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.location = Location.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.symbolId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Hover {
    return {
      markdown: isSet(object.markdown) ? globalThis.String(object.markdown) : "",
      location: isSet(object.location) ? Location.fromJSON(object.location) : undefined,
      symbolId: isSet(object.symbolId)
        ? globalThis.String(object.symbolId)
        : isSet(object.symbol_id)
        ? globalThis.String(object.symbol_id)
        : "",
    };
  },

  toJSON(message: Hover): unknown {
    const obj: any = {};
    if (message.markdown !== "") {
      obj.markdown = message.markdown;
    }
    if (message.location !== undefined) {
      obj.location = Location.toJSON(message.location);
    }
    if (message.symbolId !== "") {
      obj.symbolId = message.symbolId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Hover>, I>>(base?: I): Hover {
    return Hover.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Hover>, I>>(object: I): Hover {
    const message = createBaseHover();
    message.markdown = object.markdown ?? "";
    message.location = (object.location !== undefined && object.location !== null)
      ? Location.fromPartial(object.location)
      : undefined;
    message.symbolId = object.symbolId ?? "";
    return message;
  },
};

function createBaseCompletionItem(): CompletionItem {
  return { label: "", kind: "", detail: "", documentation: "", symbolId: "" };
}

export const CompletionItem: MessageFns<CompletionItem> = {
  encode(message: CompletionItem, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.label !== "") {
      writer.uint32(10).string(message.label);
    }
    if (message.kind !== "") {
      writer.uint32(18).string(message.kind);
    }
    if (message.detail !== "") {
      writer.uint32(26).string(message.detail);
    }
    if (message.documentation !== "") {
      writer.uint32(34).string(message.documentation);
    }
    if (message.symbolId !== "") {
      writer.uint32(42).string(message.symbolId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CompletionItem {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCompletionItem();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.label = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.kind = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.detail = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.documentation = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.symbolId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CompletionItem {
    return {
      label: isSet(object.label) ? globalThis.String(object.label) : "",
      kind: isSet(object.kind) ? globalThis.String(object.kind) : "",
      detail: isSet(object.detail) ? globalThis.String(object.detail) : "",
      documentation: isSet(object.documentation) ? globalThis.String(object.documentation) : "",
      symbolId: isSet(object.symbolId)
        ? globalThis.String(object.symbolId)
        : isSet(object.symbol_id)
        ? globalThis.String(object.symbol_id)
        : "",
    };
  },

  toJSON(message: CompletionItem): unknown {
    const obj: any = {};
    if (message.label !== "") {
      obj.label = message.label;
    }
    if (message.kind !== "") {
      obj.kind = message.kind;
    }
    if (message.detail !== "") {
      obj.detail = message.detail;
    }
    if (message.documentation !== "") {
      obj.documentation = message.documentation;
    }
    if (message.symbolId !== "") {
      obj.symbolId = message.symbolId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CompletionItem>, I>>(base?: I): CompletionItem {
    return CompletionItem.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CompletionItem>, I>>(object: I): CompletionItem {
    const message = createBaseCompletionItem();
    message.label = object.label ?? "";
    message.kind = object.kind ?? "";
    message.detail = object.detail ?? "";
    message.documentation = object.documentation ?? "";
    message.symbolId = object.symbolId ?? "";
    return message;
  },
};